aceita Felipe

+++ src/main/java/com/ecommerce/loja/repository/ProdutoRepository.java
@@ -3,5 +3,5 @@ package com.ecommerce.loja.repository;
 import com.ecommerce.loja.model.Produto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProdutoRepository extends JpaRepository<Produto, Long> {
+public interface ProdutoRepository extends JpaRepository<Produto, Integer> {
 }
